+++ src/Controller/UserController.php
@@ -23,55 +23,48 @@ use Symfony\Component\String\Slugger\SluggerInterface;
  */
 class UserController extends AbstractController
 {
-    /**
-     * @Route("/", name="user_index", methods={"GET"})
-     */
-    public function index(UserRepository $userRepository): Response
-    {
-        return $this->render('user/articles.html.twig', [
-            'users' => $userRepository->findAll(),
-        ]);
-    }
 
     /**
-     * @Route("/new", name="user_new", methods={"GET","POST"})
+     * @Route("/register", name="user_new", methods={"GET","POST"})
      */
     public function new(Request $request, UserPasswordEncoderInterface $encoder): Response
     {
-        if (!$this->getUser()) {
-            $user = new User();
-            $formUser = $this->createForm(RegistrationType::class, $user);
-            $formUser->handleRequest($request);
+        $user = new User();
+        $formUser = $this->createForm(RegistrationType::class, $user);
+        $formUser->handleRequest($request);
 
-            if ($formUser->isSubmitted() && $formUser->isValid()) {
-                $entityManager = $this->getDoctrine()->getManager();
-                $hash = $encoder->encodePassword($user,$user->getPassword());
-                $user->setPassword($hash);
-                $entityManager->persist($user);
-                $entityManager->flush();
-                return $this->redirectToRoute('app_login');
-            }
-
-            return $this->render('user/new.html.twig', [
-                'user' => $user,
-                'formUser' => $formUser->createView(),
-            ]);
+        if ($formUser->isSubmitted() && $formUser->isValid()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $hash = $encoder->encodePassword($user,$user->getPassword());
+            $user->setPassword($hash);
+            $entityManager->persist($user);
+            $entityManager->flush();
+            return $this->redirectToRoute('app_login');
         }
-        return $this->redirectToRoute('article_index');
-    }
 
+        return $this->render('user/register.html.twig', [
+            'user' => $user,
+            'formUser' => $formUser->createView(),
+        ]);
+    }
 
     /**
      * @Route("/{username}", name="user_show", methods={"GET", "POST"})
      */
     public function show(User $user, Request $request,ArticleRepository $articleRepository): Response
     {
-        $articles = $articleRepository->findBy(['author'=>$user->getId()], ['published'=>'DESC']);
-        if ($this->getUser()) {
+        $articles = $articleRepository->findBy(['author' => $user->getId()], ['published' => 'DESC']);
+        $viewParameters = [
+            'controller_name' => 'UserController',
+            'user' => $user,
+            'articles' => $articles
+        ];
+
+        if ($this->getUser() && $this->getUser()->getId() != $user->getId()) {
             $reportUser = new ReportUser();
-            $reportForm = $this->createForm(ReportUserType::class, $reportUser);
-            $reportForm->handleRequest($request);
-            if ($reportForm->isSubmitted() && $reportForm->isValid()) {
+            $formReport = $this->createForm(ReportUserType::class, $reportUser);
+            $formReport->handleRequest($request);
+            if ($formReport->isSubmitted() && $formReport->isValid()) {
                 $reportUser->setAuthor($this->getUser());
                 $reportUser->setTarget($user);
                 $reportUser->setCreated(new \DateTime('NOW'));
@@ -80,14 +73,14 @@ class UserController extends AbstractController
                 $entityManager->flush();
                 return $this->redirectToRoute('user_show', ['username' => $user->getUsername()]);
             }
+            $viewParameters['formReport'] = $formReport->createView();
 
-            $followForm = $this->createForm(UserFollowType::class, $this->getUser());
-            $followForm->handleRequest($request);
-            if ($followForm->isSubmitted() && $followForm->isValid()) {
+            $formFollowButton = $this->createForm(UserFollowType::class, $this->getUser());
+            $formFollowButton->handleRequest($request);
+            if ($formFollowButton->isSubmitted() && $formFollowButton->isValid()) {
                 if (!$this->getUser()->isFollowing($user)) {
                     $this->getUser()->addFollow($user);
-                }
-                else {
+                } else {
                     $this->getUser()->removeFollow($user);
                 }
                 $entityManager = $this->getDoctrine()->getManager();
@@ -96,65 +89,63 @@ class UserController extends AbstractController
                 $entityManager->flush();
                 return $this->redirectToRoute('user_show', ['username' => $user->getUsername()]);
             }
-
-            return $this->render('user/show.html.twig', [
-                'user' => $user,
-                'articles'=>$articles,
-                'reportForm' => $reportForm->createView(),
-                'followForm' => $followForm->createView()
-            ]);
+            $viewParameters['formFollowButton'] = $formFollowButton->createView();
         }
-        return $this->render('user/show.html.twig', [
-            'user' => $user
-        ]);
+
+        return $this->render('user/show.html.twig', $viewParameters);
     }
 
     /**
-     * @Route("/{id}/edit", name="user_edit", methods={"GET","POST", "DELETE"})
+     * @Route("/{username}/edit", name="user_edit", methods={"GET","POST", "DELETE"})
      */
     public function edit(Request $request, User $user, SluggerInterface $slugger): Response
     {
-        $formUser = $this->createForm(UserEditType::class, $user);
-        $formUser->handleRequest($request);
-        if (file_exists($user->getProfilePicture())) {
-            $user->setProfilePicture(new File($user->getProfilePicture()));
-        }
-        if (file_exists($user->getCoverPicture())) {
-            $user->setCoverPicture(new File($user->getCoverPicture()));
-        }
-
-        if ($formUser->isSubmitted() && $formUser->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            $profilePicture = $formUser->get('profile_picture')->getData();
-            if ($profilePicture) {
-                $profilePictureName = $slugger->slug($user->getUsername())."-profile.".$profilePicture->guessExtension();
-                try {
-                    $profilePicture->move($this->getParameter('user_images_directory'), $profilePictureName);
-                    $user->setProfilePicture($profilePictureName);
-                } catch (FileException $e) {}
+        if ($this->getUser() && $this->getUser()->hasAccess($user)) {
+            $formUser = $this->createForm(UserEditType::class, $user);
+            $formUser->handleRequest($request);
+            if (file_exists($user->getProfilePicture())) {
+                $user->setProfilePicture(new File($user->getProfilePicture()));
+            }
+            if (file_exists($user->getCoverPicture())) {
+                $user->setCoverPicture(new File($user->getCoverPicture()));
             }
 
-            $coverPicture = $formUser->get('cover_picture')->getData();
-            if ($coverPicture) {
-                $coverPictureName = $slugger->slug($user->getUsername())."-cover.".$coverPicture->guessExtension();
-                try {
-                    $coverPicture->move($this->getParameter('user_images_directory'), $coverPictureName);
-                    $user->setCoverPicture($coverPictureName);
-                } catch (FileException $e) {}
+            if ($formUser->isSubmitted() && $formUser->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+
+                $profilePicture = $formUser->get('profile_picture')->getData();
+                if ($profilePicture) {
+                    $profilePictureName = $slugger->slug($user->getUsername()) . "-profile." . $profilePicture->guessExtension();
+                    try {
+                        $profilePicture->move($this->getParameter('user_images_directory'), $profilePictureName);
+                        $user->setProfilePicture($profilePictureName);
+                    } catch (FileException $e) {
+                    }
+                }
+
+                $coverPicture = $formUser->get('cover_picture')->getData();
+                if ($coverPicture) {
+                    $coverPictureName = $slugger->slug($user->getUsername()) . "-cover." . $coverPicture->guessExtension();
+                    try {
+                        $coverPicture->move($this->getParameter('user_images_directory'), $coverPictureName);
+                        $user->setCoverPicture($coverPictureName);
+                    } catch (FileException $e) {
+                    }
+                }
+
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($user);
+                $entityManager->flush();
+
+                return $this->redirectToRoute('user_show', ['username' => $user->getUsername()]);
             }
 
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($user);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('user_show', ['username' => $user->getUsername()]);
+            return $this->render('user/edit.html.twig', [
+                'user' => $user,
+                'formUser' => $formUser->createView(),
+            ]);
         }
-
-        return $this->render('user/edit.html.twig', [
-            'user' => $user,
-            'formUser' => $formUser->createView(),
-        ]);
+        return $this->redirectToRoute('user_show', ['username' => $user->getUsername()]);
     }
 
     /**
@@ -163,15 +154,18 @@ class UserController extends AbstractController
     public function delete(Request $request, User $user): Response
     {
         $currentUser = $this->getUser();
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            if($user->getId() == $currentUser->getId()) {
-                $this->get('security.token_storage')->setToken(null);
+        if ($this->getUser()->hasAccess($user)) {
+            if (!$user->isAdmin() || ($user->isAdmin() && $this->getUser()->isSuperAdmin())) {
+                if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                    if ($user->getId() == $currentUser->getId()) {
+                        $this->get('security.token_storage')->setToken(null);
+                    }
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($user);
+                    $entityManager->flush();
+                }
             }
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($user);
-            $entityManager->flush();
         }
-
         return $this->redirectToRoute('admin_users');
     }
 }
